:zap: Prefill required data when asset is selected
